requirements.txt
ipykernel
python-dotenv
pandas
numpy
scikit-learn
openai
streamlit
fastapi
uvicorn